% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carpeDIAMS-package.R
\docType{package}
\name{carpeDIAMS-package}
\alias{carpeDIAMS}
\alias{carpeDIAMS-package}
\title{carpeDIAMS: DIA to pseudo-DDA spectra}
\description{
Turns DIA (bbCID) acquisitions into pseudo-DDA spectra by correlating MS1
elution with MS2 fragment elution across many samples. Produces an MGF
that downstream library/annotation tools consume.
}
\author{
\strong{Maintainer}: Johan Lassen \email{johan@example.com}

Authors:
\itemize{
  \item Johan Lassen \email{johan@example.com}
}

}
\keyword{internal}
