% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_1D_upgma_grouping_cpp}
\alias{fast_1D_upgma_grouping_cpp}
\title{Fast 1D UPGMA grouping}
\usage{
fast_1D_upgma_grouping_cpp(x, maxDiff)
}
\arguments{
\item{maxDiff}{numeric scalar}

\item{mz}{numeric vector}
}
\value{
integer vector of group ids
}
\description{
Fast 1D UPGMA grouping
}
